+++ packages/plugin-vue/src/main.ts
@@ -162,7 +162,7 @@ export async function transformMain(
   if (options.sourceMap) {
     if (scriptMap && templateMap) {
       // if the template is inlined into the main module (indicated by the presence
-      // of templateMap, we need to concatenate the two source maps.
+      // of templateMap), we need to concatenate the two source maps.
 
       const gen = fromMap(
         // version property of result.map is declared as string
@@ -213,15 +213,21 @@ export async function transformMain(
 
   // handle TS transpilation
   let resolvedCode = output.join('\n')
+  const lang = descriptor.scriptSetup?.lang || descriptor.script?.lang
+
   if (
-    (descriptor.script?.lang === 'ts' ||
-      descriptor.scriptSetup?.lang === 'ts') &&
+    lang &&
+    /tsx?$/.test(lang) &&
     !descriptor.script?.src // only normal script can have src
   ) {
     const { code, map } = await transformWithEsbuild(
       resolvedCode,
       filename,
-      { loader: 'ts', sourcemap: options.sourceMap },
+      {
+        loader: 'ts',
+        target: 'esnext',
+        sourcemap: options.sourceMap
+      },
       resolvedMap
     )
     resolvedCode = code