+++ packages/vite/src/node/optimizer/esbuildDepPlugin.ts
@@ -1,5 +1,4 @@
 import path from 'node:path'
-import { promises as fs } from 'node:fs'
 import type { ImportKind, Plugin } from 'esbuild'
 import { KNOWN_ASSET_TYPES } from '../constants'
 import { getDepOptimizationConfig } from '..'
@@ -8,7 +7,6 @@ import {
   flattenId,
   isBuiltin,
   isExternalUrl,
-  isRunningWithYarnPnp,
   moduleListContains,
   normalizePath
 } from '../utils'
@@ -164,8 +162,7 @@ export function esbuildDepPlugin(
         const flatId = flattenId(id)
         if (flatId in qualified) {
           return {
-            path: flatId,
-            namespace: 'dep'
+            path: qualified[flatId]
           }
         }
       }
@@ -181,7 +178,7 @@ export function esbuildDepPlugin(
           }
 
           // ensure esbuild uses our resolved entries
-          let entry: { path: string; namespace: string } | undefined
+          let entry: { path: string } | undefined
           // if this is an entry, return entry namespace resolve result
           if (!importer) {
             if ((entry = resolveEntry(id))) return entry
@@ -199,48 +196,6 @@ export function esbuildDepPlugin(
           }
         }
       )
-
-      // For entry files, we'll read it ourselves and construct a proxy module
-      // to retain the entry's raw id instead of file path so that esbuild
-      // outputs desired output file structure.
-      // It is necessary to do the re-exporting to separate the virtual proxy
-      // module from the actual module since the actual module may get
-      // referenced via relative imports - if we don't separate the proxy and
-      // the actual module, esbuild will create duplicated copies of the same
-      // module!
-      const root = path.resolve(config.root)
-      build.onLoad({ filter: /.*/, namespace: 'dep' }, ({ path: id }) => {
-        const entryFile = qualified[id]
-
-        let relativePath = normalizePath(path.relative(root, entryFile))
-        if (
-          !relativePath.startsWith('./') &&
-          !relativePath.startsWith('../') &&
-          relativePath !== '.'
-        ) {
-          relativePath = `./${relativePath}`
-        }
-
-        let contents = ''
-        const { hasImports, exports, hasReExports } = exportsData[id]
-        if (!hasImports && !exports.length) {
-          // cjs
-          contents += `export default require("${relativePath}");`
-        } else {
-          if (exports.includes('default')) {
-            contents += `import d from "${relativePath}";export default d;`
-          }
-          if (hasReExports || exports.length > 1 || exports[0] !== 'default') {
-            contents += `\nexport * from "${relativePath}"`
-          }
-        }
-
-        return {
-          loader: 'js',
-          contents,
-          resolveDir: root
-        }
-      })
 
       build.onLoad(
         { filter: /.*/, namespace: 'browser-external' },
@@ -276,7 +231,7 @@ module.exports = Object.create(new Proxy({}, {
       key !== 'constructor' &&
       key !== 'splice'
     ) {
-      throw new Error(\`Module "${path}" has been externalized for browser compatibility. Cannot access "${path}.\${key}" in client code.\`)
+      console.warn(\`Module "${path}" has been externalized for browser compatibility. Cannot access "${path}.\${key}" in client code.\`)
     }
   }
 }))`
@@ -300,30 +255,6 @@ module.exports = Object.create(new Proxy({}, {
           }
         }
       )
-
-      // yarn 2 pnp compat
-      if (isRunningWithYarnPnp) {
-        build.onResolve(
-          { filter: /.*/ },
-          async ({ path: id, importer, kind, resolveDir, namespace }) => {
-            const resolved = await resolve(
-              id,
-              importer,
-              kind,
-              // pass along resolveDir for entries
-              namespace === 'dep' ? resolveDir : undefined
-            )
-            if (resolved) {
-              return resolveResult(id, resolved)
-            }
-          }
-        )
-
-        build.onLoad({ filter: /.*/ }, async (args) => ({
-          contents: await fs.readFile(args.path),
-          loader: 'default'
-        }))
-      }
     }
   }
 }