+++ packages/vite/src/node/optimizer/index.ts
@@ -66,6 +66,8 @@ export interface DepsOptimizer {
   resetRegisteredIds: () => void
   ensureFirstRun: () => void
 
+  close: () => Promise<void>
+
   options: DepOptimizationOptions
 }
 
@@ -911,12 +913,22 @@ function esbuildOutputFromId(
   id: string,
   cacheDirOutputPath: string
 ): any {
+  const cwd = process.cwd()
   const flatId = flattenId(id) + '.js'
-  return outputs[
-    normalizePath(
-      path.relative(process.cwd(), path.join(cacheDirOutputPath, flatId))
-    )
-  ]
+  const normalizedOutputPath = normalizePath(
+    path.relative(cwd, path.join(cacheDirOutputPath, flatId))
+  )
+  const output = outputs[normalizedOutputPath]
+  if (output) {
+    return output
+  }
+  // If the root dir was symlinked, esbuild could return output keys as `../cwd/`
+  // Normalize keys to support this case too
+  for (const [key, value] of Object.entries(outputs)) {
+    if (normalizePath(path.relative(cwd, key)) === normalizedOutputPath) {
+      return value
+    }
+  }
 }
 
 export async function extractExportsData(
@@ -1030,7 +1042,12 @@ function isSingleDefaultExport(exports: readonly string[]) {
   return exports.length === 1 && exports[0] === 'default'
 }
 
-const lockfileFormats = ['package-lock.json', 'yarn.lock', 'pnpm-lock.yaml']
+const lockfileFormats = [
+  'package-lock.json',
+  'yarn.lock',
+  'pnpm-lock.yaml',
+  'bun.lockb'
+]
 
 export function getDepHash(config: ResolvedConfig, ssr: boolean): string {
   let content = lookupFile(config.root, lockfileFormats) || ''