+++ playground/hmr/vite.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from 'vite'
+import type { Plugin } from 'vite'
 
 export default defineConfig({
   experimental: {
@@ -19,6 +20,36 @@ export default defineConfig({
           client.send('custom:remote-add-result', { result: a + b })
         })
       }
-    }
+    },
+    virtualPlugin()
   ]
 })
+
+function virtualPlugin(): Plugin {
+  let num = 0
+  return {
+    name: 'virtual-file',
+    resolveId(id) {
+      if (id === 'virtual:file') {
+        return '\0virtual:file'
+      }
+    },
+    load(id) {
+      if (id === '\0virtual:file') {
+        return `\
+import { virtual as _virtual } from "/importedVirtual.js";
+export const virtual = _virtual + '${num}';`
+      }
+    },
+    configureServer(server) {
+      server.ws.on('virtual:increment', async () => {
+        const mod = await server.moduleGraph.getModuleByUrl('\0virtual:file')
+        console.log(mod)
+        if (mod) {
+          num++
+          server.reloadModule(mod)
+        }
+      })
+    }
+  }
+}